requirements.txt
bottle
icalendar
python-dotenv
requests
